move revert reason script
