--- core/scripts/ccip/revert-reason/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/smartcontractkit/chainlink/core/scripts/ccip/revert-reason/command"
-)
-
-func main() {
-	command.Execute()
-}
