+++ deployment/ccip/scripts/revert-reason/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/smartcontractkit/chainlink/deployment/ccip/scripts/revert-reason/command"
+)
+
+func main() {
+	command.Execute()
+}
